requirements.txt
questionary>=1.10.0
rich>=10.12.0
pyyaml>=6.0
python-nmap>=0.7.1
pymetasploit3>=1.0.3
requests>=2.26.0
asyncio>=3.4.3
torch>=2.0.0
numpy>=1.21.0